first pass at mean in numpyarraycolumngroupby
