+++ meerkat/columns/numpy_column.py
@@ -17,6 +17,7 @@ from yaml.representer import Representer
 from meerkat.block.abstract import BlockView
 from meerkat.block.numpy_block import NumpyBlock
 from meerkat.columns.abstract import AbstractColumn
+from meerkat.ops.group_by_objects import NumPyArrayGroupBy
 from meerkat.writers.concat_writer import ConcatWriter
 
 Representer.add_representer(abc.ABCMeta, Representer.represent_name)
@@ -193,6 +194,8 @@ class NumpyArrayColumn(
             # can only create a 1-D series
             return super().to_pandas()
 
+    
+
     @classmethod
     def from_npy(
         cls,
@@ -210,3 +213,6 @@ class NumpyArrayColumn(
             encoding=encoding,
         )
         return cls(data)
+
+    def to_group_by(self, indices) -> NumPyArrayGroupBy:
+        return NumPyArrayGroupBy(indices, self.data)
